+++ java/java-impl/src/com/intellij/codeInsight/completion/AllClassesGetter.java
@@ -1,18 +1,4 @@
-/*
- * Copyright 2000-2009 JetBrains s.r.o.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2000-2021 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
 package com.intellij.codeInsight.completion;
 
 import com.intellij.codeInsight.CodeInsightUtilCore;
@@ -22,32 +8,22 @@ import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.editor.RangeMarker;
 import com.intellij.openapi.progress.ProgressManager;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.util.registry.Registry;
-import com.intellij.openapi.util.text.StringUtil;
 import com.intellij.psi.*;
 import com.intellij.psi.impl.search.AllClassesSearchExecutor;
 import com.intellij.psi.impl.source.PostprocessReformattingAspect;
-import com.intellij.psi.impl.source.tree.java.PsiReferenceExpressionImpl;
 import com.intellij.psi.search.GlobalSearchScope;
 import com.intellij.util.Consumer;
 import com.intellij.util.IncorrectOperationException;
 import com.intellij.util.Processor;
-import gnu.trove.THashSet;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.Set;
 
-/**
- * Created by IntelliJ IDEA.
- * User: ik
- * Date: 02.12.2003
- * Time: 16:49:25
- * To change this template use Options | File Templates.
- */
-public class AllClassesGetter {
-  private static final Logger LOG = Logger.getInstance("#com.intellij.codeInsight.completion.AllClassesGetter");
-  public static final InsertHandler<JavaPsiClassReferenceElement> TRY_SHORTENING = new InsertHandler<JavaPsiClassReferenceElement>() {
+public final class AllClassesGetter {
+  private static final Logger LOG = Logger.getInstance(AllClassesGetter.class);
+  public static final InsertHandler<JavaPsiClassReferenceElement> TRY_SHORTENING = new InsertHandler<>() {
 
     private void _handleInsert(final InsertionContext context, final JavaPsiClassReferenceElement item) {
       final Editor editor = context.getEditor();
@@ -89,7 +65,8 @@ public class AllClassesGetter {
         }
         else if (psiClass.isValid()) {
           try {
-            context.setTailOffset(psiReference.getRangeInElement().getEndOffset() + psiReference.getElement().getTextRange().getStartOffset());
+            context
+              .setTailOffset(psiReference.getRangeInElement().getEndOffset() + psiReference.getElement().getTextRange().getStartOffset());
             final PsiElement newUnderlying = psiReference.bindToElement(psiClass);
             if (newUnderlying != null) {
               final PsiElement psiElement = CodeInsightUtilCore.forcePsiPostprocessAndRestoreElement(newUnderlying);
@@ -108,7 +85,7 @@ public class AllClassesGetter {
           }
         }
       }
-      if (toDelete.isValid()) {
+      if (toDelete != null && toDelete.isValid()) {
         document.deleteString(toDelete.getStartOffset(), toDelete.getEndOffset());
         context.setTailOffset(toDelete.getStartOffset());
       }
@@ -119,96 +96,45 @@ public class AllClassesGetter {
     }
 
     @Override
-    public void handleInsert(final InsertionContext context, final JavaPsiClassReferenceElement item) {
+    public void handleInsert(@NotNull final InsertionContext context, @NotNull final JavaPsiClassReferenceElement item) {
       _handleInsert(context, item);
       item.getTailType().processTail(context.getEditor(), context.getEditor().getCaretModel().getOffset());
     }
-
   };
 
-  public static final InsertHandler<JavaPsiClassReferenceElement> INSERT_FQN = new InsertHandler<JavaPsiClassReferenceElement>() {
-    @Override
-    public void handleInsert(InsertionContext context, JavaPsiClassReferenceElement item) {
-      final String qName = item.getQualifiedName();
-      if (qName != null) {
-        int start = context.getTailOffset() - 1;
-        while (start >= 0) {
-          final char ch = context.getDocument().getCharsSequence().charAt(start);
-          if (!Character.isJavaIdentifierPart(ch) && ch != '.') break;
-          start--;
-        }
-        context.getDocument().replaceString(start + 1, context.getTailOffset(), qName);
-        LOG.assertTrue(context.getTailOffset() >= 0);
-      }
+  public static final InsertHandler<JavaPsiClassReferenceElement> INSERT_FQN = (context, item) -> {
+    final String qName = item.getQualifiedName();
+    if (qName != null) {
+      int start = JavaCompletionUtil.findQualifiedNameStart(context);
+      context.getDocument().replaceString(start, context.getTailOffset(), qName);
+      LOG.assertTrue(context.getTailOffset() >= 0);
     }
   };
 
   public static void processJavaClasses(@NotNull final CompletionParameters parameters,
                                         @NotNull final PrefixMatcher prefixMatcher,
                                         final boolean filterByScope,
-                                        @NotNull final Consumer<PsiClass> consumer) {
+                                        @NotNull final Consumer<? super PsiClass> consumer) {
     final PsiElement context = parameters.getPosition();
     final Project project = context.getProject();
     final GlobalSearchScope scope = filterByScope ? context.getContainingFile().getResolveScope() : GlobalSearchScope.allScope(project);
 
-    Processor<PsiClass> processor = new Processor<PsiClass>() {
-      final Set<String> qNames = new THashSet<String>();
-      final boolean pkgContext = JavaCompletionUtil.inSomePackage(context);
-      final String packagePrefix = getPackagePrefix(context, parameters.getOffset());
-
-      @Override
-      public boolean process(PsiClass psiClass) {
-        if (parameters.getInvocationCount() < 2) {
-          if (PsiReferenceExpressionImpl.seemsScrambled(psiClass)) {
-            return true;
-          }
-          if (!StringUtil.isCapitalized(psiClass.getName()) && !Registry.is("ide.completion.show.lower.case.classes")) {
-            return true;
-          }
-        }
-
-        assert psiClass != null;
-        if (isAcceptableInContext(context, psiClass, filterByScope, pkgContext)) {
-          String qName = psiClass.getQualifiedName();
-          if (qName != null && qName.startsWith(packagePrefix) && qNames.add(qName)) {
-            consumer.consume(psiClass);
-          }
-        }
-        return true;
-      }
-    };
-    processJavaClasses(prefixMatcher, project, scope, processor);
+    processJavaClasses(prefixMatcher, project, scope, new LimitedAccessibleClassPreprocessor(parameters, filterByScope, c->{consumer.consume(c); return true;}));
   }
 
   public static void processJavaClasses(@NotNull final PrefixMatcher prefixMatcher,
                                         @NotNull Project project,
                                         @NotNull GlobalSearchScope scope,
-                                        @NotNull Processor<PsiClass> processor) {
-    final Set<String> names = new THashSet<String>(10000);
-    AllClassesSearchExecutor.processClassNames(project, scope, new Consumer<String>() {
-      @Override
-      public void consume(String s) {
-        if (prefixMatcher.prefixMatches(s)) {
-          names.add(s);
-        }
+                                        @NotNull Processor<? super PsiClass> processor) {
+    final Set<String> names = new HashSet<>(10000);
+    AllClassesSearchExecutor.processClassNames(project, scope, s -> {
+      if (prefixMatcher.prefixMatches(s)) {
+        names.add(s);
       }
+      return true;
     });
-    LinkedHashSet<String> sorted = CompletionUtil.sortMatching(prefixMatcher, names);
+    LinkedHashSet<String> sorted = prefixMatcher.sortMatching(names);
     AllClassesSearchExecutor.processClassesByNames(project, scope, sorted, processor);
-  }
-
-
-  private static String getPackagePrefix(final PsiElement context, final int offset) {
-    final CharSequence fileText = context.getContainingFile().getViewProvider().getContents();
-    int i = offset - 1;
-    while (i >= 0) {
-      final char c = fileText.charAt(i);
-      if (!Character.isJavaIdentifierPart(c) && c != '.') break;
-      i--;
-    }
-    String prefix = fileText.subSequence(i + 1, offset).toString();
-    final int j = prefix.lastIndexOf('.');
-    return j > 0 ? prefix.substring(0, j) : "";
   }
 
   public static boolean isAcceptableInContext(@NotNull final PsiElement context,
